Add Support for the new BSI Tokens and a cleaner experience with Legacy URLs (#48)
* Unnecessary help text removal

* Remove additional options

Leaving in these "Overrides" that take over the default states for
legacy support perhaps?  I'm not sure yet.

* Add option to prefer remediation scan if available

* Create FodApiFilterList class

this is to just make it so we have less string concat going on.

* Hook up remediation preferred option

* Clean up remediation scan and service calls

* Remove redundant url building

* Move this help to the global options section

* Move remediation to Entitlement options

* Add in override options

* Hook up the overrides.

If the values are set in the BSI Token, these will write over the value
when on, but take no affect when not selected.  I need to document that.

* I need to stop messing this up...

* Fix issue with wrong TechnologyBlahblahblah was used

* shorten this call

* AH did it again!